+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix bug with copying project specific Catalogues between Projects when committing a Cohort
 - Remove Tree filters from UI
 - Fix issue with Web File Downloader not releasing files after download
+- Fix Extraction primary Key case sensitivity issue
 
 ## [9.1.2] - Unreleased
 - Automatically fetch user settings from previous versions of RDMP when installing the latest version